# swagger.yaml
openapi: 3.0.0
info: 
  version: 1.0.0
  title: ShareTunes
  description: SpotifyAPI Information
  contact:
    name: Group 11
...
paths:
  /fetch/artistv2/:
    get:
      summary: Searchs for an artist and returns the artist with the exact matching name
      tags:
        - fetchV2
      parameters:
        - in: query
          name: artist
          required: true
          schema:
              type: string
          description: Search for artist by exact name
          example: Drake
      responses:
        '200':
          description: artist found
          content:
            'application/json':
                schema:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    image:
                      type: string
        '500':
          description: artist not found
        '401':
          description: Bad or expired token
        '403':
          description: Bad OAuth request
        '429':
          description: The app has exceeded its rate limits
  /fetch/artistsv2/:
    get:
      summary: Searchs for an artist and returns the artists with matching names
      tags:
        - fetchV2
      parameters:
        - in: query
          name: artist
          required: true
          schema:
              type: string
          description: Search artist's name
          example: Cardi B
      responses:
          '200':
            description: artist found
            content:
              'application/json':
                schema:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      id:
                        type: string
                      image:
                        type: string
          '500':
            description: artist not found
          '401':
            description: Bad or expired token
          '403':
            description: Bad OAuth request
          '429':
            description: The app has exceeded its rate limits
  /fetch/trackv2/:
    get:
      summary: Searches for a track by name and returns the tracks with matching names.
      tags:
        - fetchV2
      parameters:
        - in: query
          name: track
          required: true
          schema:
              type: string
          description: Searches track name
          example: Under My Umbrella
      responses:
        '200':
          description: artist found
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    artist:
                      type: string
                    preview:
                      type: string
                    url_link:
                      type: string
        '500':
          description: song not found
        '401':
          description: Bad or expired token
        '403':
          description: Bad OAuth request
        '429':
          description: The app has exceeded its rate limits
  /fetch/recsv2/:
    get:
      summary: Gets playlist reccomendations based on seeds.
      tags:
        - fetchV2
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: object
            properties:
              seed_artists:
                type: array
                items:
                  type: string
              seed_genres:
                type: array
                items:
                  type: string
              seed_tracks:
                  type: array
                  items:
                    type: string
          description: Gets recommendations based on seeds. It only accepts 5 seeds in total. seed_artists takes in artistIDs and seed_tracks takes in trackIDs
          example:
            seed_artist: [4kYSro6naA4h99UJvo89HB, 3TVXtAsR1Inumwj472S9r4]
            seed_genres: [hip-hop, r-n-b]
            seed_tracks: [3A2yGHWIzmGEIolwonU69h]
      responses:
        '200':
          description: Returns an array of tracks
          content:
            'application/json':
              schema:                 
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    artist:
                      type: string
                    preview:
                      type: string
                    url_link:
                      type: string  
        '500':
            description: Error getting recommendations
        '401':
            description: Bad or expired token
        '403':
            description: Bad OAuth request
        '429':
            description: The app has exceeded its rate limits
  /fetch/recspipev2/:
    get:
      summary: GET RECCOMENDATIONS FROM STRINGS => Gets playlist reccomendations based on seeds. NO trackIDs or artistIDS required.
      tags:
        - fetchV2
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: object
            properties:
              artists:
                type: array
                items:
                  type: string
              genres:
                type: array
                items:
                  type: string
              tracks:
                  type: array
                  items:
                    type: string
          description: Gets recommendations based on seeds. It only accepts 5 values in total.
          example:
            artists: ["Rihanna", "Beyonce"]
            genres: ["hip-hop"]
            tracks: ["Formation", "Jumpin"]
      responses:
        '200':
          description: Returns an array of tracks
          content:
            'application/json':
              schema:                 
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    artist:
                      type: string
                    preview:
                      type: string
                    url_link:
                      type: string  
        '500':
            description: Error getting recommendations
        '401':
            description: Bad or expired token
        '403':
            description: Bad OAuth request
        '429':
            description: The app has exceeded its rate limits
...
  /fetch/artist:
    get:
      summary: SEARCH ARTIST => Searchs for an artist and returns the artist with the exact matching name
      tags:
        - fetch
      requestBody:
        description: Search for artist by exact name.
        required: true
        content:
          application/x-www-form-urlencoded::
            schema:
              type: object
              properties:
                artist:
                  type: string
              example: 
                artist: Drake
                description: Artist's name.
      responses:
        '200':
          description: artist found
          content:
            'application/json':
                schema:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    image:
                      type: string
        '500':
          description: artist not found
        '401':
          description: Bad or expired token
        '403':
          description: Bad OAuth request
        '429':
          description: The app has exceeded its rate limits
  /fetch/artists:
    get:
      summary: SEARCH ARTISTS => Searchs for an artist and returns the artists with matching names
      tags:
        - fetch
      requestBody:
        description: Search artist's name. 
        required: true
        content:
          application/x-www-form-urlencoded::
            schema:
              type: object
              properties:
                artist:
                  type: string
              example: 
                artist: Cardi B
                description: Artist's name.
      responses:
          '200':
            description: artist found
            content:
              'application/json':
                schema:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      id:
                        type: string
                      image:
                        type: string
          '500':
            description: artist not found
          '401':
            description: Bad or expired token
          '403':
            description: Bad OAuth request
          '429':
            description: The app has exceeded its rate limits
  /fetch/track:
    get:
      summary: SEARCH TRACK => Searches for a track by name and returns the tracks with matching names.
      tags:
        - fetch
      requestBody:
        description: Searches track name. 
        required: true
        content:
          application/x-www-form-urlencoded::
            schema:
              type: object
              properties:
                track:
                  type: string
              example: 
                track: Under My Umbrella
                description: Track name.
      responses:
        '200':
          description: artist found
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    artist:
                      type: string
                    preview:
                      type: string
                    url_link:
                      type: string
        '500':
          description: song not found
        '401':
          description: Bad or expired token
        '403':
          description: Bad OAuth request
        '429':
          description: The app has exceeded its rate limits
  /fetch/genres:
    get:
      summary: GET GENRES => Get Spotify's available genres
      tags:
        - fetch
      responses:
        '200':
          description: Returns an array of Spotify's available genres
          content:
            'application/json':
                schema:                 
                  type: array
                  items:
                    type: string  
        '500':
          description: Error connecting to Spotify
        '401':
          description: Bad or expired token
        '403':
          description: Bad OAuth request
        '429':
          description: The app has exceeded its rate limits
  /fetch/recs:
    get:
      summary: GET RECCOMENDATIONS => Gets playlist reccomendations based on seeds.
      tags:
        - fetch
      requestBody:
        description: Searches track name. 
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                seed_artists:
                  type: array
                  items:
                    type: string
                seed_genres:
                  type: array
                  items:
                    type: string
                seed_tracks:
                    type: array
                    items:
                      type: string
              description: Gets recommendations based on seeds. It only accepts 5 seeds in total. seed_artists takes in artistIDs and seed_tracks takes in trackIDs
              example:
                seed_artist: [4kYSro6naA4h99UJvo89HB, 3TVXtAsR1Inumwj472S9r4]
                seed_genres: [hip-hop, r-n-b]
                seed_tracks: [3A2yGHWIzmGEIolwonU69h]
      responses:
        '200':
          description: Returns an array of tracks
          content:
            'application/json':
              schema:                 
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    artist:
                      type: string
                    preview:
                      type: string
                    url_link:
                      type: string  
        '500':
            description: Error getting recommendations
        '401':
            description: Bad or expired token
        '403':
            description: Bad OAuth request
        '429':
            description: The app has exceeded its rate limits
  /fetch/recspipe:
    get:
      summary: GET RECCOMENDATIONS FROM STRINGS => Gets playlist reccomendations based on seeds. NO trackIDs or artistIDS required.
      tags:
        - fetch
      requestBody:
        description: Names of Artists, Tracks, and Genres. 
        required: true
        content:
          application/x-www-form-urlencoded::
            schema:
              type: object
              properties:
                artists:
                  type: array
                  items:
                    type: string
                genres:
                  type: array
                  items:
                    type: string
                tracks:
                    type: array
                    items:
                      type: string
              description: Gets recommendations based on seeds. It only accepts 5 values in total.
              example:
                artists: ["Rihanna", "Beyonce" ]
                genres: [hip hop]
                tracks: ["Formation", "Jumpin"]
      responses:
        '200':
          description: Returns an array of tracks
          content:
            'application/json':
              schema:                 
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    id:
                      type: string
                    artist:
                      type: string
                    preview:
                      type: string
                    url_link:
                      type: string  
        '500':
            description: Error getting recommendations
        '401':
            description: Bad or expired token
        '403':
            description: Bad OAuth request
        '429':
            description: The app has exceeded its rate limits